+++ src/components/AppHeader/index.js
@@ -1,0 +1,71 @@
+// @flow
+import * as React from 'react';
+import { css, cx } from 'react-emotion';
+import {
+  Breadcrumbs,
+  type BreadcrumbsProps,
+  ControlsPanel
+} from '../../index';
+
+const styles = {
+  header: css`
+    height: 68px;
+    background: #ffffff;
+    width: 100%;
+    display: flex;
+    padding: 0 30px;
+    box-sizing: border-box;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+  `,
+  crumbs: css`
+    display: flex;
+    flex-grow: 1;
+    flex-shrink: 1;
+    flex-direction: row;
+    flex-wrap: nowrap;
+    align-items: center;
+  `,
+  controls: css`
+    display: flex;
+    flex-basis: auto;
+    flex-grow: 0;
+    flex-shrink: 0;
+    flex-direction: row;
+    flex-wrap: nowrap;
+    align-items: center;
+    margin-left: 20px;
+  `
+}
+
+type AppHeaderProps = {
+  ...BreadcrumbsProps,
+  className?: string,
+  controls?: React.Node[]
+}
+
+export const AppHeader = (
+  {
+    appName,
+    breadcrumbs,
+    children,
+    className,
+    controls,
+    onLinkClick
+  }: AppHeaderProps
+) => (
+  <div className={cx(styles.header, className)}>
+    <Breadcrumbs
+      className={styles.crumbs}
+      items={breadcrumbs}
+      appName={appName}
+      onLinkClick={onLinkClick}
+    />
+    <ControlsPanel
+      className={styles.controls}
+      controls={controls}
+      thin
+    />
+  </div>
+);