+++ CHANGELOG.md
@@ -14,11 +14,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Input
   - InputPassword
   - TextArea
+  - Breadcrumbs
 
 - Updated icons:
 
   - IconEyeClosed
   - IconEyeOpened
+
+- Added AppHeader component.
 
 ## [0.23.0] - 2020-09-01